new functionality in working state

+++ src/com/trafficmon/EntryEvent.java
@@ -1,7 +1,7 @@
 package com.trafficmon;
 
 public class EntryEvent extends ZoneBoundaryCrossing {
-    public EntryEvent(Vehicle vehicleRegistration,int time) {
+    public EntryEvent(Vehicle vehicleRegistration,double time) {
         super(vehicleRegistration,time);
     }
 }
