+++ src/com/trafficmon/ExitEvent.java
@@ -1,7 +1,7 @@
 package com.trafficmon;
 
 public class ExitEvent extends ZoneBoundaryCrossing {
-    public ExitEvent(Vehicle vehicle,int time) {
+    public ExitEvent(Vehicle vehicle,double time) {
         super(vehicle,time);
     }
 }
